运动控制卡/轴/参数表结构与服务全面升级
- 完善 motion_card、motion_axis、motion_axis_config 等实体结构，新增显示名、驱动键、分组、单位、描述、时间戳等字段，类型更规范，注释更清晰
- 服务接口与实现补充唯一性校验、字段归一化、自动索引、时间戳赋值，参数保存支持类型/范围/分组自动校验
- 种子数据与配置装配同步适配新字段，聚合对象字段补全
- 前端 ViewModel 支持新字段的增删改查与热重载
- 文档补充 motion_axis_config 设计原则，强调参数表不承载 UI/枚举定义
- 显著提升运动配置的规范性、唯一性和可扩展性

+++ AM.DBService/Services/Motion/App/AxisConfigAppService.cs
@@ -406,7 +406,7 @@ namespace AM.DBService.Services.Motion.App
         /// <summary>
         /// 创建数据库参数对象。
         /// </summary>
-        private static MotionAxisConfigEntity CreateRow(int logicalAxis, string axisName, string paramName, string displayName, string valueType, double value)
+        private static MotionAxisConfigEntity CreateRow(short logicalAxis, string axisName, string paramName, string displayName, string valueType, double value)
         {
             return new MotionAxisConfigEntity
             {
